pyproject.toml
[project]
name = "ai-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipython>=9.5.0",
    "langchain>=0.3.27",
    "langchain-openai>=0.3.33",
    "langgraph>=0.6.7",
    "matplotlib>=3.10.6",
    "openai>=1.107.3",
    "pandas>=2.3.2",
    "python-dotenv>=1.1.1",
    "rich>=14.1.0",
]
